Make the script aware of target architecture
Add the ability to parse the processor architecture from the config
file. The user can override the architecture with the -a. Additionally,
if the user wants to use the -p option to print the recommendations
without specifying a kernel config file, the -a option can be used to
print the recommendations that correspond to the specificied
architecture.

Some recommendations are architecture specific so we need to warn the
user if they're checking a kernel config for an architecture that
doesn't have any architecture specific recommendations.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>