Make the script aware of target architecture and kernel version
Add the ability to parse the processor architecture and kernel version
from the config file. The user can override the architecture and version
with the -a and -k options. Additionally, if the user wants to use the
-p option to print the recommendations without specifying a kernel
config file, the -a and -k options can be used to print the
recommendations that correspond to the specificied architecture and
version.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>